# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductsController_listProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductsController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_getProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_updateProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/active": {
      "patch": {
        "operationId": "ProductsController_setActive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProductActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/inventory/adjustments": {
      "post": {
        "operationId": "InventoryController_createAdjustment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdjustmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNotificationReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}/send-whatsapp": {
      "post": {
        "operationId": "NotificationsController_sendWhatsApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationWhatsAppDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/purchases": {
      "get": {
        "operationId": "PurchasesController_listPurchases",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Purchases"
        ]
      },
      "post": {
        "operationId": "PurchasesController_createPurchase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Purchases"
        ]
      }
    },
...
    "/packages": {
      "get": {
        "operationId": "PackagesController_getAllPackages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Packages"
        ]
      },
      "post": {
        "operationId": "PackagesController_createPackage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Packages"
        ]
      }
    },
    "/packages/{id}": {
      "patch": {
        "operationId": "PackagesController_updatePackage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Packages"
        ]
      }
    },
...
    "/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_getReservation",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      },
      "patch": {
        "operationId": "ReservationsController_updateReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
...
    "/blocked-slots": {
      "get": {
        "operationId": "BlockedSlotsController_listBlockedSlots",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BlockedSlots"
        ]
      },
      "post": {
        "operationId": "BlockedSlotsController_createBlockedSlot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockedSlotDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "BlockedSlots"
        ]
      }
    },
    "/blocked-slots/{id}": {
      "patch": {
        "operationId": "BlockedSlotsController_updateBlockedSlot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockedSlotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BlockedSlots"
        ]
      },
      "delete": {
        "operationId": "BlockedSlotsController_deleteBlockedSlot",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BlockedSlots"
        ]
      }
    },
    "/public/reservations/{token}": {
      "get": {
        "operationId": "PublicReservationsController_getByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicReservations"
        ]
      },
      "patch": {
        "operationId": "PublicReservationsController_updateByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePublicReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicReservations"
        ]
      }
    }
  },
  "info": {
    "title": "MAGIC CITY API",
    "description": "API demo funcional para manejo de eventos",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc_session"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
      "SetProductActiveDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAdjustmentDto": {
        "type": "object",
        "properties": {}
      },
...
      "MarkNotificationReadDto": {
        "type": "object",
        "properties": {}
      },
      "SendNotificationWhatsAppDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePurchaseDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePackageDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePackageDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateReservationDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBlockedSlotDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBlockedSlotDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePublicReservationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}